Initial split of commands, using NewCmd* pattern.
